Removed jest & cleaned up the index a bit
