+++ src/acceptEmailInput/acceptEmailInput.js
